requirements.txt
playwright
playwright-stealth>=2.0.0
requests
google-genai
python-dotenv